Move releaseinfo out of control-plane-operator
The releaseinfo package is used by the ignition server and hypershift operators.
Neither should have a dependency on control-plane-operator which will eventually
move out of this repository.

This commit moves the releaseinfo package to the root of the hypershift
repository.
