+++ hypershift-operator/main.go
@@ -23,9 +23,9 @@ import (
 
 	"github.com/go-logr/logr"
 	hyperapi "github.com/openshift/hypershift/api"
-	"github.com/openshift/hypershift/control-plane-operator/releaseinfo"
 	"github.com/openshift/hypershift/hypershift-operator/controllers/hostedcluster"
 	"github.com/openshift/hypershift/hypershift-operator/controllers/nodepool"
+	"github.com/openshift/hypershift/releaseinfo"
 	"github.com/spf13/cobra"
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/exporters/otlp"
