+++ releaseinfo/fixtures/4.8-image-references.json
