+++ releaseinfo/fixtures/4.8-installer-coreos-bootimages.yaml
