+++ releaseinfo/static_provider.go
